requirements.txt
streamlit==1.38.0
pandas==2.2.2
numpy==1.26.4
faiss-cpu==1.8.0.post1
sentence-transformers==2.7.0
pymupdf==1.24.9
sqlalchemy==2.0.32
pydantic==2.9.2
python-dotenv==1.0.1
reportlab==4.2.2

